refactor file structure for new seed
