+++ src/newLaunch/baseStage.scss
