+++ src/newLaunch/seed/newSeed.html
@@ -1,5 +1,5 @@
 <template>
-  <require from="../resources/elements/circledNumber/circledNumber.html"></require>
+  <require from="resources/elements/circledNumber/circledNumber.html"></require>
 
   <div class="newSeedContainer ${currentStage.name}">
     <div class="sideBarDropDownContainer">
