+++ src/newLaunch/seed/seedConfig.ts
