+++ src/newLaunch/seed/stage4.html
@@ -1,5 +1,5 @@
 <template>
-  <require from="./tokenLabel.html"></require>
+  <require from="../tokenLabel.html"></require>
   <div class="page animated-page stageContainer stage4 au-animate ">
     <div class="infoContainer">
       <div class="heading heading2 title">SEED Details</div>
