+++ src/newLaunch/seed/stage5.ts
@@ -1,4 +1,4 @@
-import { BaseStage } from "newSeed/baseStage";
+import { BaseStage } from "newLaunch/baseStage";
 import { Utils } from "services/utils";
 
 export class Stage5 extends BaseStage {
