+++ src/newLaunch/seed/stage6.ts
@@ -1,5 +1,5 @@
 import { autoinject } from "aurelia-framework";
-import { BaseStage } from "newSeed/baseStage";
+import { BaseStage } from "newLaunch/baseStage";
 
 @autoinject
 export class Stage6 extends BaseStage {
