+++ src/newLaunch/tokenLabel.html
